Add bogus source jar.
Sonatype now requires a source jar for every artifact
being uploaded. gwt-servlet did not have a source jar
in the past and for that reason upload to maven central
fails.
This patch creates a bogus jar file with no sources in it
for gwt-servlet, thus we pass validation for central.
We need to properly resolve this in the build to produce
an actual source file for gwt-servlet.

Change-Id: I366f8d90a2084f4b4b29b8bab6de0fd14baa0b64