Use gwt-user-sources.jar as gwt-servlet-sources.jar for Maven.
Sonatype requires a sources jar, and gwt-servlet doesn't include sources
(on purpose) and our script to create sources jars just extracts the
sources from the jar, so it doesn't create one for gwt-servlet. Our
workaround is thus to just reuse the source jar for gwt-user as
gwt-servlet is basically a subset of gwt-user.

Change-Id: I2b6b529a730c92cd1a6b25cb3609eb6d08744714