Update patch set 1
Patch Set 1:

Ok if we add a flag, then lets make sure to make it preceded by 'X', otherwise people complain when we remove it.

Also in the demo, I saw the <class_name>.<method_name>. Most of the time (not always), the class name is redundant as the frame already shows the file name and kind of distracts. I think it is better to just have the method name and covering the common case.

Patch-set: 1