Update patch set 1
Patch Set 1:

We could have NONE|METHOD|CLASS|FULL in the compiler but we will need to pick one as the default for Super Dev Mode.

Another use case I thought of is that if you have sourcemaps off (to debug a JavaScript issue), the Java filename won't appear so you won't have the class.

Patch-set: 1
Label: Verified=0