Update patch set 5
Patch Set 5:

> I'm really wondering if adding the sources to gwt-servlet.jar is a
 > good idea.

Hm. I had assumed the plan was for .class files to go n gwt-servlet.jar (as today AFAIU), and .java files to only go in a separate gwt-servlet-sources.jar, which is the standard Maven/Java convention.

Doesn't seem like copying the .java files into gwt-servlet.jar (which is admittedly GWT client conventions) is worth it, unless this is merely because it's easiest to do this in the current build systems.

 > How about just using gwt-user-sources.jar in lib-gwt.sh as the
 > SOURCES_JAR for gwt-servlet?

That is admittedly what I've done in the past for "damn it I need *some* sources jar to pacify Eclipse", but I always assumed that surely the gwt-user and gwt-servlet contents were somehow different, such that my hack was only a heuristic and so invalid for some potentially large set of .class files in gwt-servlet.

Patch-set: 5
Reviewer: Gerrit User 5025 <5025@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0