Introduce PerformanceCounter as a way of getting compiler metrics
Any compiler class that has a TreeLogger can increment a counter,
with the result available to code that embeds Super Dev Mode.

Started with one counter for DeclaredTypesInModule as proof of
concept.

Also, fix error-handling bugs in Super Dev Mode:
 - JobEvent VALID_TAG regular expression was incorrect.
 - JobEvent was exposing rebased ImmutableList class
 - If a Job failed with a RuntimeException that happened outside the
compiler, the stack trace was not being logged and the job would
never complete.

Change-Id: Icb6f7e457c6d9b3b7de341a21b18b2ea0d505177