Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

I know this is a quick fix, but I don't thing is a good idea to produce two files referring to the same module differing so much. 

servlet.jar has 3734 classes, 540 undocumented in user-sources.jar
user.src.jar has 3615 classes, 421 which are not in servlet.jar

I think a best option should be produce a servlet-sources.jar with exactly the same files which there are in servlet.jar, but that implies maintaining .src files of the rebased libraries in svn to extract those sources.

The real problem is when the servlet.jar is in your eclipse classpath debug is pretty upsetting when coming to a class included in servlet  without sources.

Patch-set: 1
Reviewer: Gerrit User 5055 <5055@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+1