Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > I know this is a quick fix, but I don't thing is a good idea to
 > produce two files referring to the same module differing so much.
 > 
 > servlet.jar has 3734 classes, 540 undocumented in user-sources.jar

Are they documented anywhere else? (gwt-dev)
gwt-servlet includes a bit of gwt-dev, but they're only a bunch of internal classes (and thirdparty classes).

 > I think a best option should be produce a servlet-sources.jar with
 > exactly the same files which there are in servlet.jar, but that
 > implies maintaining .src files of the rebased libraries in svn to
 > extract those sources.

It also means either packaging that -sources.jar in the GWT SDK (we do that for the requestfactory-*.jar so why not) or reworking entirely the way we deploy to Maven.
But gwt-servlet is already way too big, including many classes that can't even be used in a plain JVM, and the GWT SDK is huge too, so adding even more Mb for no practical added value (just attach gwt-user.jar as the source jar for gwt-servlet.jar in Eclipse) is debatable.

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0