Update patch set 2
Patch Set 2:

> In general, we should avoid any static state in generators. What are we caching in GssResourceGenerator right now?

We are caching the ASTs (in a "independant permutation" state) by using as key the list of the urls of gss file , so that we can can skip the parsing and some gss compilers passes if the gss source files have not been modified.
I mainly did that because it was done like this in the current CssResourceGenerator and I thought it was needed for sdm. But now that we have incremental compilation, it's maybe not needed anymore.
Also I've dug in the git history to find the reason why they've introduced this static cache and the reason was to avoid repeated parsing of the same source material when you uses UiBindes's implicit stylesheets (UiBinderGenerator needs to parse the stylesheet and extract the style class names in order to generate the CssResource interface). But with GSS, we don't reuse this cache anymore (UiBinderGenerator instantiates a GssParser and runs one pass to extract the class names).
So if we should avoid any static state in generators, I propose to remove this static cache from the GssResourceGenerator because it is not useful if we don't use it in UiBinderGenerator. What do you guys think ?

Patch-set: 2