Update patch set 3
Patch Set 3: -Code-Review

(1 comment)

Just one nit.

Thomas, the idea is that the UiBinderParser should not deal with the default value (decide what is the default value if the attribute is not set) and pass the Boolean as is to the class that will consume it.
It's ok here.

Patch-set: 3
Label: Code-Review=0