Enable GSS to  CSSResourceTest.
The goal of this change is to duplicate the tests for the current
CssResourceGenerator and enable GSS and the converter in order to
test the backward compatibility. GSS produces a more compact css
and some assertions have been adapted.
When GSS will be enabled by default, the existing CssResourceTest
class will be removed in favor of GwtCssResourceTestWithGss.

Change-Id: Ie4b28d1c89cd44b0adaf3a30c3f87191d1827774