Disable testReplaceItemNoEvent for HtmlUnit platform
When running tests with HtmlUnit, when testReplaceItemNoEvent adds a new history
token with the fire event set to false, the event is not fired during the test,
but some time after, and testEmptyHistoryTokens sees it.
For some reason and randomly testEmptyHistoryTokens does not fail, probably
because of the order how the tests are run changes, or depending on how busy the
server is.

Change-Id: I52ee77616770e42e2a33a71ffe4125ad80940bcb