Prevent browser default behaviour on arrow up/down in SuggestBox.
When pressing the arrow up/down keys in a SuggestBox on Chrome, the cursor
jumps to the beginning/end of the text box. This is unexpected if the
suggestion display is shown, and the arrow up/down keys are used to navigate
the suggestion menu. Firefox does not seem to have this behaviour. This
change proposes to prevent default browser behaviour.

Change-Id: Icf311f65106ebd62b612a20fb9c383706e18aa33