Update patch set 3
Patch Set 3:

It turns out that the main() method we're calling already blocks until the code server is started, so we just need to wait for it to finish.

I did a performance test using "hello" and Super Dev Mode starts in about 10 seconds, and we save 1 second with an async API. I think a 10% performance gain is significant if it holds up with larger apps and enough to justify a slightly more complex API. I'll try showcase next.

Patch-set: 3