Update patch set 2
Patch Set 2:

about the work to reuse caches in https://gwt-review.googlesource.com/#/c/9450/ i put it on hold since it didn't look like it would make it in time for the 2.7 release and since it's actually a little bit of a dangerous change.

if there are any bugs in either the persistentunitcache or the minimalrebuildcache then they will become more impactful when caching are being preserved and reused for long periods of time. so my thinking was to let things settle a bit before putting this in.

but, i'm expecting to start working on c/9450 again late this week or early next week

Patch-set: 2