Update patch set 9
Patch Set 9:

> Should this be part of 2.7 release?

Well, it's not strictly necessary, but recommendable. 
Running ant build in maven projects fails, and this patch just notice the user about he has to use maven.
There is no way without this patch to run superdevmode in showcase, dynatable or mail projects.
When I did this patch for some reason mobilewebapp was failing because of the gae version, but right now it does not fails, it could be a maven repo issue, but anyway we should update gae version since it shows a message to upgrade.

So I vote to have this into 2.7.

Patch-set: 9
Label: Verified=0