Attempt to fix synthetic keyboard events for IE11
The patch falls back to KeyboardEvent constructor per the
W3C spec if Firefox specific method doesn't exist.

I am not sure if this is going to work in IE11 but hopefully
eventually it should work. As this method was already throwing
an exception before so it is not a step back.


Bug: issue 8463
Change-Id: I8230e542c7a6ec74418d62f6c4b3aadf41de59e1
Review-Link: https://gwt-review.googlesource.com/#/c/10202/