Update patch set 1
Patch Set 1:

> I'm not sure we can take this into the GWT proper.
 > OracleJDK/OpenJDK is licensed under GPLv2+Classpath Exception, GWT
 > is licensed under Apache.
 > 
 > I know, it sucks, it's mostly API definitions, but in order to
 > avoid trouble, we need to avoid copying their files directly.
 > 
 > It doesn't take that much effort (I did it for like 5 of them a
 > while ago for testing) and we can land things incrementally and
 > just convert a few at a time.

IMHO that is nonsense but ok I am not a lawyer. 80% are given by API definition. The remaining 20% are logical implementations.

One thing I would change for GWT though is to use Goktugs InternalPrecondition class instead of Objects.requireNonNull() so that users can tell GWT to remove all these non-null checks in the future. That would affect quite some default methods.

I could do that in this CL or create a new one in which I manually create files on disk. But don't expect the default method implementations to be different from this CL except the requireNonNull() stuff.

Patch-set: 1