Fix for HeaderPanel hidden widgets when set again.
This fixes the bug that on a HeaderPanel when a header, content or footer widget is set again it is hidden. This was caused because the container of these widgets is hidden when the previous widget is removed (in the add). The widget was made visible, but before the add, therefor rendering it useless. This fix only makes the containers visible if the new widget passed is not null.

Bug: issue 7037
Change-Id: I2b41970c37112db257089dc648534de1f257fc00