Update patch set 1
Patch Set 1:

Hmm not sure if that is a good starting point to not fix bugs because someone might depend on it. If it would be a really high profile bug I can imagine the situation but here it's a minor issue. I would think that the goal should be to create bug free software?
The problem is triggered when in code the widget is set more then once. The first time it doesn't trigger (because remove is not called). So users of the widget already have to set it again to make the widget visible. The workaround normaly done is to make the widget visible in the users own code, (Which very annoying, not to mention the unexpected behaviour when someone uses the widget only to wonder why it is hidden). The side effect of fixing this bug is probably that the code the user wrote (making the widget visible again) will become obsolete. The use case that would depend of this unexpected behavior would be the user expects the widget to be hidden and wants to make it visible later. I don't think that is a common use case. At some point the widget has to be made visible again. And if the user depends on it if this bug is fixed the user will find out directly because the widget becomes visible directly when added, opposite of the current unexpected behavior were it is hidden when added a second time.

Patch-set: 1
Label: Verified=0