Update patch set 4
Patch Set 4:

Actually, trying to make TypeMapper just return a JMethodCall expression leads to yuckiness all over the place. For example, there's a number of calls throughout GenJsAST that assume they're dealing with literals and calling getRuntimeTypeLiteral(). These then delegate to JjsUtil.translateLiteral which hope to get back a JsExpression.

If I try to pass through a JMethodCall, then when convertJavaLiteral() is called, I'll have to unpack the JMethodCall and convert it on the fly to a JsExpression. That seems far uglier than introducing a new literal, unless you can suggest a simpler way to shoehorn this in.

For example, when the defineClass() arguments are being setup, how would you take the result of typeMapper and turn it into a jsExpression?

Patch-set: 4
Label: Verified=0