Update patch set 2
Patch Set 2:

1. It's impossible to make that happen, that's how subtyping works. If you @IterateAsArray, you are essentially saying your type is asserting random access. It's not much different than subtypes of Set not guaranteeding ordering, and subtypes of List guaranteeing order.

The behavior of subtypes will be changed no matter what as soon as you upcast to the supertype.


Basically, what you're saying is:

for(T x : arrayList)  should pick up the semantic change, but for(T x : subtypeOfArrayList) shouldn't, but if you declare a method like this:

void process(ArrayList<T> al) {
   for (T x : al) { ...}
}

and pass subtypeOfArrayList to it, you're going to get this behavior anyway. 

Not propagating the contract down is more incorrect. It makes upcasting a type operate differently.

Also, without subtyping logic, it would require all JsInterop classes that inherit JsArrayLike to redeclare @IterateAsArray in order to work. But this is a compiler hint that most people shouldn't have to use, it's more for the SDK or library authors.

2. On point #2, the purpose of this isn't to avoid precondition checks, it's to avoid the allocation of an Iterator, double property lookup, and NoSuchElementException. You might remove the latter (althought it might break some code)

You can run benchmarks, but I'm pretty sure in a type loop, that

for(i = 0, max=arraylist.array.length; i < max; i++) {
   x=arraylist.array[i];
}

is going to beat

while(i < iterator.arraylist.array.length) {
   x = iterator.arraylist.array[i++];
}

And that's assuming you remove the check for no-such element exeception and everything inlines perfectly.

For-each on an integer loop is guaranteed not to exceed the end of the arraylist.

Patch-set: 2