Update patch set 2
Patch Set 2:

On the first remark, my point is, if you cannot prove that the upper bound of the type is not ArrayList, it is not safe. But is true that it makes it impractical for JsArrayLike. However that probably means we should probably only use it for ArrayLike.

On the second issue; after preconditions are removed, we need to benchmark to see how the end code looks like and if there are any low hanging fruits that can give us the similar performance.

We will get some results soon so I think it is good to wait little bit more.

Patch-set: 2