Update patch set 2
Patch Set 2:

I'd like to see an example that shows the bad behavior to convince me.

Using @IterateAsArray is a contract. It says your type (and it's subtypes) have O(1) randomized access. 

If you extend a random access class with one that isn't (for example, subtype ArrayList and give it linked-list performance), your code is broken in many more ways than just @IterateAsArray.

This annotation is much safer than the SideEffect annotation (which if used incorrect leads to broken code, not slow code), and safer than substituting Integers for unboxed types. In the worse case, if something goes wrong, you convert an O(1) lookup into an O(n) lookup, but you can only do that by writing code that violates it's own super-type contract.

All this patch does is the following (using the ArrayList example)

1) if the collection variable of the for-each loop is provably ArrayList, use integer loop
2) if the collection variable's supertype is provably ArrayList, use integer loop

This is absolutely, 100% safe as far as I can tell. This patch NEVER causes a collection that's a super-type of ArrayList or sibling of ArrayList (like LinkedList) to get this behavior.

it simply says for(T x : C) gets an integer loop if and only if C is an IterateAsArray type, or it's parent is an IterateAsArray type. No type that is not the descendant of an IterateAsArray type will ever get this optimization.

If you have written a class C that is O(1) get() method, read the JavaDoc for @IterateAsArray, and then went ahead and implemented a subtype D of C whose's get() method is O(n), then you have written code that's going to have terrible surprising performance regardless of whether this annotation exists or not.

The JavaDoc contract for ArrayList is "The size, isEmpty, get, set, iterator, and listIterator operations run in constant time". @IterateAsArray has the same constraints for implementators that ArrayList's JavaDoc imposes on people who subtype ArrayList.

Patch-set: 2