Update patch set 2
Patch Set 2:

I think you misunderstood me. When I say it is not safe, I'm not saying it is not safe because it might end up being slow. ArrayList already marked with RandomAccess so basically any subclass promises to have good performance on indexed lookups.

The 'safeness' issue is; you are breaking the java contract of for-each loop that calls iterator() on classes that are not in our control. A very trivial example for that is there will be no way for Guava to have concurrent modification protection because iterator is no longer called.

This is a tradeoff and it is not a trivial one so we need to put all pros and cons in place before making a decision.

Patch-set: 2