Update patch set 2
Patch Set 2:

Your right, I was thinking of something else. Well, if you want to say this is a JsInterop only annotation, fine.

I would argue that CME checks should optimize away in production mode and should only be thrown in checked mode, but I don't want to bikeshed on this. We could make this optimization prod-mode only.

But in any case, this annotation is absolutely needed to make for-each stylistically work for Javascript collections without creating expensive iterator wrappers.

Patch-set: 2