Update patch set 2
Patch Set 2:

> Your right, I was thinking of something else. Well, if you want to
 > say this is a JsInterop only annotation, fine.
 > 
 > I would argue that CME checks should optimize away in production
 > mode and should only be thrown in checked mode, but I don't want to
 > bikeshed on this. We could make this optimization prod-mode only.
 > 
 > But in any case, this annotation is absolutely needed to make
 > for-each stylistically work for Javascript collections without
 > creating expensive iterator wrappers.

I think ConcurrentModificationException is a distraction, my main point is the effects of breaking the java contract. IsArrayLike is a different story and we can define our own contract so I'm not worried about that part.

Going back to my original argument, I'm not saying we shouldn't have it. Let's wait little bit more to better justify it or perhaps make it opt-in etc.

Patch-set: 2