i18n: remove timestamp and versions from generated files
The timestamps result in lots of churn in the diffs each time
we regenerate the files. Every file appears to be touched
even if it's just fixing a minor bug. Removing this
info now should make diffs that are smaller and more
meaningful in the future.

I'll commit the generated files as a separate patch.

Change-Id: I0c5f13047d1bc5ce3eec951a93f0f3a3621287b9