Drop effect on DataTransfer object.
This patch updates the data transfer object to implements the
drop effect.

For the dragenter and dragover events, the dropEffect will be initialized
based on what action the user is requesting. How this is determined is
platform specific, but typically the user can press modifier keys to
adjust which action is desired. Within an event handler for the dragenter
and dragover events, the dropEffect should be modified if the action the
user is requesting is not the one that is desired.

For dragstart, drag, and dragleave events, the dropEffect is initialized
to "none". Any value assigned to the dropEffect will be set, but the value
isn't used for anything.

For the drop and dragend events, the dropEffect will be initialized to the
action that was desired, which will be the value that the dropEffect had
after the last dragenter or dragover event.

Change-Id: I09e75c35357d081c40f2827dd7b2e995b08f1a93