Update patch set 1
Patch Set 1:

I should have added more explanation here as goktug originally suggested. Belatedly, I'm doing it now:

For a very long time there has been a (rarely-exposed) but in JUnitShell when switching back and forth between prod mode and dev mode test compiles.

The issue is that DevMode was written to *not* overwrite existing Foo.nocache.js files in an attempt to play nice with production application compiles in the same directory.

This behavior made it so that if one ran a prod mode test and then a dev mode test, the dev mode test execution would not overwrite the JUnit.nocache.js file, and it only *happened* to still work most of the time because by default the JUnit.nocache.js file created by the prod mode test compile was compatible. But if one happened to add a -userAgents Blah argument to the prod mode test launch, an incompatible JUnit.nocache.js file would be created and when used in the dev mode test execution it would fail.

The reason this became highly visible in release 2.7 was because JUnitShell was modified to have some known specific default values for the -userAgents property where previously it had been unrestricted by default.

So to summarize, DevMode's desire to not overwrite existing Foo.nocache.js files might be the right thing to do in interactive use but is the wrong thing to do for test executions.

We talked about it internally and there were three options considered:
1. Add a 2nd hack on top of the existing DevMode specific linker hack, to specialize it to act differently in test runs.
2. Leave things alone and allow DevMode based tests to start breaking (DevMode is deprecated and will hopefully soon be deleted).
2. Leave things alone and allow DevMode based interactive launches start breaking in the rare circumstance that a user actually depends on this "do not overwrite" behavior. (DevMode is deprecated and will hopefully soon be deleted).

We chose option #3 since we thought users depending on this "do not overwrite" behavior in interactive launches would be rarer than users depending on DevMode based tests.

Patch-set: 1
Reviewer: Gerrit User 5410 <5410@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0