Update patch set 2
Patch Set 2:

Based on this debug output, it looks like "TRY" is the best currency to use and "TL" is an variant. But the code is just looping over fields and the last one wins. Maybe these fields are reached in a different order than before?

     [java] addCurrencyExtras for en
     [java] For path: //ldml/numbers/currencies/currency[@type="TRY"]/symbol
     [java] Got symbol 'TRY' for TRY
     [java] For path: //ldml/numbers/currencies/currency[@type="TRY"]/symbol[@alt="variant"]
     [java] Got symbol 'TL' for TRY
     [java] For path: //ldml/numbers/currencies/currency[@type="TRY"]/symbol[@alt="narrow"]
     [java] Got symbol '₺' for TRY
     [java] ignored symbol

I haven't figured out where the data is coming from yet.

Patch-set: 2