Update patch set 1
Patch Set 1:

I think I figured it out. SimpleXMLSource.iterator() iterates over the keys of SimpleXmlSource.xpath_value.keySet(). This is a ConcurrentHashMap, so the order is not guaranteed to be anything in particular. Perhaps it has changed between different Java versions?

In any case, for the output to be deterministic, LocaleData.addCurrencyEntries() should not be depending on the order of the iterator. Instead of taking the last Currency object it sees, it should have some consistent policy for choosing between them.

I'm not sure what that order should be though. I could sort the xpaths and hope for the best? This would at least put longer xpaths after shorter ones, so they would tend to win.

Patch-set: 1
Reviewer: Gerrit User 5010 <5010@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0