Workaround for suprious exceptions in -compileReport.
Passing the -compileReport flag makes the compiler produce readable
html reports from the xml reports produced by soyc; these xml reports
are parsed with SAX and in some cases SAX callbacks receive
corrupted values that cause an exception in the report producing
code.

Surprisingly, wrapping the GZIPInput stream into a BufferedInputStream
solved the issue.

This code that produces the reports shouldeither be rewritten to use
the original soyc data (not parsing the soyc output) or replaced by
a web app.

Change-Id: I9ced50ac0fee1ea2374b178a4ef34ecf96ac6c60