Update patch set 5
Patch Set 5: Code-Review+1

(2 comments)

Look good. The only thing other than my comment related to warning is; I'm not sure if the verb "restrict" used here is a good one as it can mean something else in 'property' context. For example binding properties are 'restricted' to some values but configuration properties are "set" to the values. In gwt.xml files, we use the verb 'set' consistently instead. Anyway as John and Roberto are the maintainers here they should decide if this matters or not to them.

Patch-set: 5
Label: Code-Review=+1