Update patch set 2
Patch Set 2:

> Shouldn't we fix something in either the gwt-maven-plugin or GWT
 > proper wrt those broken links?

In my opinion the compileReport should check whether the compilerMetrics  is given and show the link in case.  Or maybe we could enable metrics always if report is set, but the problem is that compiler Metrics is experimental. I vote to remove the experimental flag and enableMetrics.

Forcing compilerMetrics in gwt-maven is an option, but we force the user to use an experimental feature.

Patch-set: 2