Update patch set 2
Patch Set 2:

I considered avoiding loading modules from Devmode.
I agree it is a good idea to do that but we should avoid reintroducing thread (I guess you are not suggesting that). The execution scope of of CodeSever main is not clear, it may start accessing more static state anytime so in general it is not a good idea.

Patch-set: 2
Label: Verified=0