Add support for JsonArray in JRE environment with an initial capacity.
Both, the Java and the browser implementation of an JsonArray can be
optimized when the final size of it is known (even when just
approximated).
Unfortunately an ArrayList with capacity 5 is different from a JsonArray
with length 5. While also the ArrayList uses a 5 element array in that
case, the length of 5 is not observable, but the size() method will
return the amount of elements actually inserted.
To offer user still the optimization opportunity, the initialCapacity
has been added to the JreJsonFactory only.

Bug: issue 9058
Change-Id: I8f6039698728e371c02996cfb34e99fa56d96abc