Add support for JsonArray with an initial capacity
Both, the Java and the browser implementation of an JsonArray can be
optimized when the final size of it is known (even when just
approximated). In Java ArrayList offers a constructor with an int
parameter. In GWT JavaScriptObject.createArray also recommends to pass
in a size parameter when the size is known.
This change exposes the initialCapacity parameter to JsonFactory and
related classes.

Bug: issue 9058
Change-Id: I8f6039698728e371c02996cfb34e99fa56d96abc