Update patch set 3
Patch Set 3: Code-Review-1

Actually, there's a problem with this approach, a discrepancy between JreJsonArray and JsJsonArray: creating an Array with a given "capacity" in JS sets the array's 'length' to the given value (the Array appears to be filled with 'undefined'), contrary to to an ArrayList in Java.

That means we should either:

 * use the JS behavior, and fill the JreJsonArray with nulls, but that means we should make sure that JsJsonArray behaves correctly in the presence of 'undefined', and I believe today they're not supported (because 'undefined' doesn't exist in JSON)
 * use the Java behavior, and just use the same zero-arg JavaScriptObject.createArray() in the JS implementation.
 * only have the factory method with initial capacity in the JreJsonFactory, and neither on JsonFactory or JsJsonFactory. That means code that wants to take advantage of the optimization would have to know in which environment it'll run.

I'd lean towards the second or third option here. If you can go with the third option for your use-case, then let's just do that, it'd be a first step towards the second option if we feel the need for it in the future.

Patch-set: 3
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=-1