Enhances TypeTightener with ability to lock into exact types.
Being able to lock to an exact type helps us to mark even more
methods as non-polymorphic which provides new optimization
opportunities.

Earlier, for example, if there was any reference to LinkedHashMap
in the app, we were not able to apply method specializations in
HashMap because all specialized method was potentially polymorphic.
With the new optimization if the variable is only assigned to
HashMap instance, now we potentially identify that and specialize
the methods. Same is true for any optimizations that exploit the
fact that a method doesn't require polymorphic dispatch
(like inlining).

The implementation introduces a new JReferenceType called JExactType
that extends from JNotNullType. JExactType indicates the type cannot
be a subclass (in analogy to JNotNullType's cannot be null).
For example, if JExactType refers to HashMap, the type cannot be any
subtype but the HashMap itself.
To seed the compiler, the JConstructor returns the JExactType
of the encapsulating type, then it propagates to rest of the system
via TypeTightener through type tightenings mechanisms.

Change-Id: Iad95138d336bbda66bbdbcc6d007b1b722d1c604