Update patch set 3
Patch Set 3:

It is a bit unfortunate that there is basically this API difference.
While in fact the initial capacity causes an array of nulls to be created, as in JavaScript, the List API hides this fact again and will tel you size == 0.
Also it is a bit unfortunate that the Suggested usage of JsObject.createArray() exists but cannot be used here in a generic way. I will use the third option you suggested, but in that case there cannot be any test for that, because that is basically not observable.

Patch-set: 3
Reviewer: Gerrit User 8105 <8105@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0