Pass the native stack to console.error for better SDM experience.
Instead of attempting to produce a nicely formatted string based
on the Java-version of the stack (previously done in TextLogFormatter),
we log the native JS stacks directly to the console.

This allows browsers like Chrome to the source map to map/link the
obfuscated function names of each stack frame back to the Java
line number.

Because JS does not know about nested exceptions, we have to manually
unroll causes.

Heavily based on Jens's code for native logging for the GWT uncaught exception
handler:

https://gist.github.com/jnehlmeier/cddbc476fd330b1d4999

Bug: 9046 (only the logger portion)
Change-Id: I0de93f2df955d65144ade6ba88e5bac6b8ef5678