Fixes a test broken with 5272935761ede8e9f505e21ba8dbdf7198475234
I forget to include this fix in the original patch.

This is due to removal of special casing of JParameter in
TypeTightener. As unused parameter were pruned anyway,
TypeTightener was not tightening them when there was no
assignments, now it is tightened.
Even though it shouldn't effect the end result, the test
was implicitly relying on those not being optimized.

Change-Id: I2048e9399c713e274d309ec8a41756df156145b6