Update patch set 12
Patch Set 12:

We don't need to support older versions (before 8) on an ongoing basis, but keeping the change behind the gwt.rpc.version flag temporarily will help those of us on trunk migrate apps to the newer protocol.

Since this patch was pretty close to done, perhaps we should just try to land it? Then follow up with a patch to change the default to version 8, and then drop the old protocol versions and simplify the code.

Patch-set: 12
Label: Verified=0