Syncs incremental and opt Enum.valueOf(Foo.class, "A") handling.
Previous to control-flow based incremental link time pruning if a class
referenced FooEnum in anyway (whether a particular instance or the 
enum's class literal) then the entire enum class would be kept. And 
since the content of that class would reference the synthetic nested 
$MAP class it would be kept as well.

But with the introduction of control-flow based pruning the synthetic 
nested $MAP class could only be kept if control flow could see that the
FooEnum.valueOf() method was called (since only that function 
references it). But if instead only Enum.valueOf(FooEnum.class) was 
called, control flow could not see that FooEnum.valueOf() was called 
because the implementation of Enum.valueOf() hides the method dispatch 
in untyped runtime JS execution.

Since it is theoretically possible for any FooEnum whose class literal 
is referenced to have its FooEnum.valueOf() method reflectively called 
at runtime through the Enum.valueOf(FooEnum.class) dispatch route then 
(in the absence of data flow analysis) to be safe one must assume that 
the FooEnum.valueOf() method is called.

Change-Id: I5ccd8d9d667be7d62efd4f7ee874195401b44f1c